moved support manager to a new component
