+++ phpstan.neon
@@ -46,14 +46,14 @@ parameters:
 
         - '#Call to an undefined static method Viserio\\Component\\Support\\Str::#'
         - '#Call to an undefined method Viserio\\Component\\View\\View::withFoo#'
-        - '#Call to an undefined method Viserio\\Component\\Support\\Tests\\Fixture\\TestConnectionManager::getName#'
         - '#Call to an undefined method Viserio\\Component\\Console\\Tests\\Command\\ClosureCommandTest::comment#'
         - '#Call to an undefined method Viserio\\Component\\Log\\Logger::getName#'
         - '#Call to an undefined method Viserio\\Component\\Mail\\Mailer::foo#'
-        - '#Call to an undefined method Viserio\\Component\\Support\\Tests\\Fixture\\TestManager::set#'
-        - '#Call to an undefined method Viserio\\Component\\Support\\Tests\\Fixture\\TestManager::get#'
-        - '#Call to an undefined method Viserio\\Component\\Support\\Tests\\Fixture\\TestConnectionManager::set#'
-        - '#Call to an undefined method Viserio\\Component\\Support\\Tests\\Fixture\\TestConnectionManager::get#'
+        - '#Call to an undefined method Viserio\\Component\\Manager\\Tests\\Fixture\\TestManager::set#'
+        - '#Call to an undefined method Viserio\\Component\\Manager\\Tests\\Fixture\\TestManager::get#'
+        - '#Call to an undefined method Viserio\\Component\\Manager\\Tests\\Fixture\\TestConnectionManager::getName#'
+        - '#Call to an undefined method Viserio\\Component\\Manager\\Tests\\Fixture\\TestConnectionManager::set#'
+        - '#Call to an undefined method Viserio\\Component\\Manager\\Tests\\Fixture\\TestConnectionManager::get#'
         - '#Call to an undefined method Viserio\\Component\\Contract\\Routing\\Router::foo#'
         - '#Cannot call method getTarget#'
         - '#Cannot call method getName#'
