+++ phpunit.xml.dist
@@ -86,6 +86,9 @@
         <testsuite name="Narrowspark Mail Component Test Suite">
             <directory>./src/Viserio/Component/Mail/Tests</directory>
         </testsuite>
+        <testsuite name="Narrowspark Manager Component Test Suite">
+            <directory>./src/Viserio/Component/Manager/Tests</directory>
+        </testsuite>
         <testsuite name="Narrowspark OptionsResolver Component Test Suite">
             <directory>./src/Viserio/Component/OptionsResolver/Tests</directory>
         </testsuite>
