+++ src/Viserio/Component/Cache/composer.json
@@ -39,8 +39,8 @@
         "cache/void-adapter": "^1.0",
         "psr/container": "^1.0",
         "viserio/contract": "self.version",
-        "viserio/options-resolver": "self.version",
-        "viserio/support": "self.version"
+        "viserio/manager": "self.version",
+        "viserio/options-resolver": "self.version"
     },
     "provide": {
         "psr/cache-implementation": "^1.0",
