+++ src/Viserio/Component/Contract/Cache/Manager.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 namespace Viserio\Component\Contract\Cache;
 
-use Viserio\Component\Contract\Support\Manager as BaseManager;
+use Viserio\Component\Contract\Manager\Manager as BaseManager;
 
 interface Manager extends BaseManager
 {
