+++ src/Viserio/Component/Contract/Manager/Exception/Exception.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Contract\Manager\Exception;
+
+/**
+ * Exception interface for all exceptions thrown by the component.
+ */
+interface Exception extends \Throwable
+{
+}
