+++ src/Viserio/Component/Contract/Manager/Exception/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Component\Contract\Manager\Exception;
+
+use InvalidArgumentException as BaseInvalidArgumentException;
+
+class InvalidArgumentException extends BaseInvalidArgumentException implements Exception
+{
+}
