--- src/Viserio/Component/Contract/Support/Connector.php
@@ -1,17 +1,0 @@
-<?php
-declare(strict_types=1);
-namespace Viserio\Component\Contract\Support;
-
-interface Connector
-{
-    /**
-     * Establish a connection.
-     *
-     * @param array $config
-     *
-     * @throws \InvalidArgumentException|\RuntimeException
-     *
-     * @return object
-     */
-    public function connect(array $config): object;
-}
