+++ src/Viserio/Component/Mail/composer.json
@@ -28,6 +28,7 @@
         "swiftmailer/swiftmailer": "^6.0",
         "symfony/polyfill-mbstring": "^1.8.0",
         "viserio/contract": "self.version",
+        "viserio/manager": "self.version",
         "viserio/options-resolver": "self.version",
         "viserio/support": "self.version"
     },
