+++ src/Viserio/Component/Manager/Tests/AbstractConnectionManagerTest.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
-namespace Viserio\Component\Support\Tests;
+namespace Viserio\Component\Manager\Tests;
 
 use Narrowspark\TestingHelper\ArrayContainer;
 use Narrowspark\TestingHelper\Phpunit\MockeryTestCase;
 use stdClass;
-use Viserio\Component\Contract\Support\Exception\InvalidArgumentException;
-use Viserio\Component\Support\Tests\Fixture\TestConnectionManager;
+use Viserio\Component\Contract\Manager\Exception\InvalidArgumentException;
+use Viserio\Component\Manager\Tests\Fixture\TestConnectionManager;
 
 /**
  * @internal
