+++ src/Viserio/Component/Session/composer.json
@@ -29,6 +29,7 @@
         "paragonie/halite": "^4.4.2",
         "symfony/polyfill-mbstring": "^1.8.0",
         "viserio/contract": "self.version",
+        "viserio/manager": "self.version",
         "viserio/options-resolver": "self.version",
         "viserio/support": "self.version"
     },
